+++ CHANGELOG.md
@@ -7,7 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-- Nothing yet.
+- Implement CLI
+- Implement AutoHotkey scripts for common actions
 
 ## [0.0.1]